+++ example/ios/RunnerTests/RunnerTests.swift
@@ -1,4 +1,5 @@
 import Flutter
+import SVGKit
 import UIKit
 import XCTest
 
@@ -10,6 +11,18 @@ import XCTest
 // See https://developer.apple.com/documentation/xctest for more information about using XCTest.
 
 class RunnerTests: XCTestCase {
+
+  func testSVGIconRenders() throws {
+    let data = Data("""
+      <svg xmlns="http://www.w3.org/2000/svg" width="10" height="10" viewBox="0 0 10 10">
+        <rect width="10" height="10" fill="red"/>
+      </svg>
+      """.utf8)
+    let svg = try XCTUnwrap(SVGKImage(data: data))
+    let image = try XCTUnwrap(svg.uiImage)
+
+    XCTAssertEqual(image.size, CGSize(width: 10, height: 10))
+  }
 
   func testGetPlatformVersion() {
     let plugin = MobileAppPrivacyPlugin()